apiBacorex/requirement.txt
flask-restful
Flask  
datetime 
uuid
Flask-SQLAlchemy
PyJWT